1.Modify the Etag flag to accommodate Baidu Cloud Object Storage. 2.Modify the contentlength type to handle files larger than 2G.

+++ src/aws/s3/writeparse.rs
@@ -3076,14 +3076,14 @@ impl TransitionStorageClassWriter {
     }
 }
 
-pub type ContentLength = i32;
+pub type ContentLength = i64;
 /// Parse `ContentLength` from XML
 pub struct ContentLengthParser;
 
 impl ContentLengthParser {
     pub fn parse_xml<T: Peek + Next>(tag_name: &str, stack: &mut T) -> Result<ContentLength, XmlParseError> {
         try!(start_element(tag_name, stack));
-        let obj = i32::from_str(try!(characters(stack)).as_ref()).unwrap();
+        let obj = i64::from_str(try!(characters(stack)).as_ref()).unwrap();
         try!(end_element(tag_name, stack));
         Ok(obj)
     }
@@ -3930,14 +3930,14 @@ impl S3ClientMessageWriter {
 
 
 
-pub type Size = i32;
+pub type Size = i64;
 /// Parse `Size` from XML
 pub struct SizeParser;
 
 impl SizeParser {
     pub fn parse_xml<T: Peek + Next>(tag_name: &str, stack: &mut T) -> Result<Size, XmlParseError> {
         try!(start_element(tag_name, stack));
-        let obj = i32::from_str(try!(characters(stack)).as_ref()).unwrap();
+        let obj = i64::from_str(try!(characters(stack)).as_ref()).unwrap();
         try!(end_element(tag_name, stack));
         Ok(obj)
     }
